ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0

# Data and ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.13.0
implicit==0.7.2

# Database
pymysql==1.1.0
sqlalchemy==2.0.30
python-dotenv==1.0.1

# Cache
redis==5.0.4

# HTTP
httpx==0.27.0

# Monitoring
sentry-sdk[fastapi]==1.44.1
